requirements.txt
# 视频生成项目依赖
requests>=2.28.0
volcenginesdkarkruntime>=1.0.0
ffmpeg-python>=0.2.0